Split AdviceContentBuilder.cs into partial classes (#334)
* Add "Powered by Performance Studio" line on landing page


* Add Darling Data favicon to web app


* Add Open Graph and Twitter Card meta tags for social sharing

Uses the Darling Data barbell logo as hero image when shared on
social media. Also adds meta description for SEO.


* Clarify OG description: in-browser, nothing to install


* Fix Rule 3 severity: CouldNotGenerateValidParallelPlan is actionable

This reason means something in the query blocks parallelism (scalar UDFs,
table variable inserts, etc.) — that's worth a Warning, not Info.


* Expand Rule 3 to cover all NonParallelPlanReason values

Adds human-readable messages for all 25 known reasons. Severity:
- Warning: actionable reasons (UDFs, cursors, table variables, remote
  queries, trace flags, hints, DML OUTPUT, writeback variables)
- Info: passive/environmental (cost below threshold, edition limits,
  memory-optimized tables, upgrade mode, index build edge cases)


* Split AdviceContentBuilder.cs into partial classes

Move-only refactor; no behavior changes. AdviceContentBuilder.cs (1,227
lines) split into 4 partials:

  NodeLinks  (243) - MakeNodeRefsClickable + ProcessInlines +
                     AddRunsWithNodeLinks + WireNodeClickHandler
  Operators  (251) - CreateOperatorLine/Group/TimingLine +
                     CreateWarningBlock
  WaitStats  (242) - CreateWaitStatLine + CreateMissingIndexImpactLine +
                     ParseWaitMs + GetWaitCategoryBrush +
                     CreateTriageSummaryCard
  Sql        ( 36) - BuildSqlHighlightedLine

Main file now 444 lines — brushes/fonts, PhysicalOperators &
SqlKeywords sets, regex constants, the 3 Build() overloads (entry
points), IsSubSectionLabel helper.

Class made `internal static partial`.

Build clean: 0 errors on PlanViewer.App.


---------
